Temporarily removing support for netstandard1.3
We've hit an issue compiling for netstandard1.3 using the latest version of the dotnet CLI (reported here: https://github.com/dotnet/cli/issues/2739). The current mechanism used to incorrectly report success during compile, but required runtime references couldn't be found with the latest dotnet CLI.  Until we have a resolution to the issue or better guidance on how to compile for DNXCore, it's safer in the short term to remove that build target.