Fixing Pull Image and fixing Create Image to use Pull.
Create image was the only way to perform a pull, but it turns out that pull image is a superset of the create image functionality.  This change adds in pull and just has create call that method.
NOTE: It also cleans up some autogenerated types (getting rid of AuthConfigParameters which was a duplicate of AuthConfig), which will require clients to update the name of that class.