Timespan for Docker classes (#207)
* Initial Implementation of TimeSpan

Started with specgen.go
Added type customizations for Config's StopTimeout, HealthConfig's
Interval, and HealthConfig's Timeout setting their values to TimeSpan?
Then added a new map of typeCustomizationKey => CSAttribute
Finally, in reflectTypeMembers, if there a map for the type and Property
name, append the attribute to the Property's list of attributes.

After this, considered several options with respect to serializing
TimeSpans as longs, and opted to write a Converter and Resolver.

TimeSpanSerializationAttribute can decoreate a Property and is given a
SerializationTarget (Seconds || Nanoseconds).

The JsonSerializer now has a TimeSpanContractResolver. The
TimeSpanContractResolver overrides only one function CreateProperty, and
uses the return value of base.CreateProperty as it's starting point. If
the Property is decorated with a TimeSpanSerializationAttribute, then it
will detect what is the target for that property (Second or Nanosecond)
and will set the converter for that property as appropriate.

Added two converters TimeSpanSecondsConverter and TimeSpanNanoseconds
for converting from/to Seconds/Nanoseconds respectively.

* Fixed Issues with Serialization

Now serializing to longs explicitly and added the converters to the
JsonConverter

* calcUsings supports usings on Property Attributes

Attributes on properties from different namespaces will be added to the
usings list

* Timespan Support via JsonConvert

Instead of using a custom ContractResolver and custom attributes,
annotate the Properties with the Newtonsoft.Json [JsonConvert]
attribute, and provide the type converter